changed measure-based key detection to note-based key detection
