+++ www/client/src/modules/Midi2MusicXML/types.ts
@@ -12,7 +12,6 @@ export type Note = {
 export type Measure = {
     notes: Note[];
     section: Section;
-    keyCandidates: number[];
 };
 
 export type Section = {
